requirements.txt
plotly
matplotlib
pandas
xlrd
openpyxl
tk
dash
numpy
pillow